Minor security improvement
Now the symbols to differentiate between filedata and plaintext are prepended after the encryption. This means there are absolutely no hints as to whether the data has been successfully decrypted or not. The only vulnerability is knowing whether it's file data or plaintext, which doesn't give any of the actual data away.